+++ Provider/FileProviderExtension+Errors.swift
@@ -6,4 +6,7 @@ enum FileProviderExtensionError: Error {
     case skipped
     case unableToOpenFile
     case s3ItemParseFailed
+    case fatal
+    case parseError
+    case fileNotFound
 }
