+++ Provider/S3Item.swift
@@ -10,9 +10,9 @@ class S3Item: NSObject, NSFileProviderItem {
     let identifier: NSFileProviderItemIdentifier
     
     private let logger = Logger(subsystem: "io.cubbit.CubbitDS3Sync.provider", category: "S3Item")
-    private let metadata: S3Item.Metadata
     private let separator = DefaultSettings.S3.delimiter
     
+    let metadata: S3Item.Metadata
     let drive: DS3Drive
     
     init(
@@ -29,6 +29,15 @@ class S3Item: NSObject, NSFileProviderItem {
         } else {
             self.identifier = identifier
         }
+    }
+    
+    init(from item: NSFileProviderItem, drive: DS3Drive) {
+        self.identifier = item.itemIdentifier
+        self.drive = drive
+        self.metadata = S3Item.Metadata(
+            lastModified: item.contentModificationDate as? Date,
+            size: (item.documentSize ?? 0) ?? 0
+        )
     }
     
     var itemIdentifier: NSFileProviderItemIdentifier {
